POWER: add PhoenixLite support to BCI
The PhoenixLite supports two different mode of operation for its charger.
One is very similar to twl6030 with only a couple of tweaks. The other
the charger control is moved to hardware and the driver becomes just
a monitor for hardware state.

This patch also corrects the uV/mV confusion in the driver. All values
are stored in mV and returned to userspace in uV to maintain alignment
with the regulator framework voltages.

Change-Id: I5920c3b042017da5ae8a72a0bb12854e2f7c342e
Signed-off-by: Graeme Gregory <gg@slimlogic.co.uk>
Signed-off-by: Volodymyr Riazantsev <v.riazantsev@ti.com>