omap: gcxxx: Add initial omap gcxxx device file.
To be used to intialize the Vivante gc-core driver for 2d hw blts on omap.

We need to also pass a common plat data struct to the driver, so we add it
to the pubic gccore driver header.  Since it needs to be shared, relocate
the header to include/linux and fix-up driver source paths.

Signed-off-by: David Sin <davidsin@ti.com>